refactor: switch backend endpoint
