+++ src/modals/SignUpModal/container/hooks/SignupModalForm.ts
@@ -10,7 +10,6 @@ export type SignupModalFormData = {
 };
 
 export const useSignupModalForm = () => {
-  // 관리 될 form 의 초기값
   const defaultValues = {
     emailId: '',
     emailDomain: '',
@@ -20,7 +19,6 @@ export const useSignupModalForm = () => {
     nickname: '',
   };
 
-  // react-hook-form이 제공해주는 form 관리 hook
   const form = useForm({
     defaultValues: {
       ...defaultValues,
