+++ src/modals/SignUpModal/state.ts
@@ -1,6 +1,3 @@
-// step들이 들어감
-// step에 따라 view를 갈아끼워넣기
-
 import { atom } from 'recoil';
 import { EStep } from './type';
