+++ src/shared/index.ts
@@ -1,4 +1,2 @@
-// 프로젝트 전체를 아우르는 공유 컴포넌트가 들어갈 폴더
-
-export * from './Button';
+export * from './Button/Button';
 export * from './Pressable/Pressable';
